Support for detached head
Changed the code to checkout detached head for new projects.

If the project is on a detached head we will simply check it out to
latest on update and user will loose any local uncomitted changes.

If the project is on a untracked branch, we do nothing and just give
user a message.

If the project is on a tracked branch we try to rebase it to remote, if
that fails, we restore the project and give user msg.

Change-Id: I86009769707154412d5ba1ada402c04b011125fe