Cache should not be created with --depth flag
--mirror flag does not work well with --depth flag on next fetch, it
fetches all the refs and then discards them, therefore taking lot of time
on each update.

Repo cache can be referenced by two projects one having depth and one
which do not, so cache should not be shallow as it serves both kind of
projects.

As it is only created once it will have no effect on subsquent update
on bots.

Change-Id: I6e66e5fae9594cc7d04f6a2049d4846c03f855f1