[attributes] Adds "attributes" XML attributes to manifest
This patch adds "attributes" attribute to projects and packages, which
will help to group different projects and packages and can be used in
git attributes in the future. When "attributes" is explictly set for a
project or package, it will not be fetched by default. These projects
can be fetched by setting optional attributes using
'jiri init -fetch-optional=att1,attr2'.

Bug: VC-154

Test: CQ
Change-Id: I40e085cd2efef0d89c82fb5254c5908ddcf3fabc