[resolve] Add optional flag to prevent floating refs.
This patch adds optional flag 'allow-floating-refs' to prevent
using a floating ref of a package. By default this flag is set to false,
which means floating refs are not allowed. This patch will help us to prevent
devs puting floating refs in manifest, which may causing merge conflicts
on bots.

Bugs: VC-158
Change-Id: I30efd502e5c5e5101af6e693e6020062e9bdf393