[attributes] Allow jiri update snapshot to override optional attrs
This change adds '-override-optional'  to 'jiri update' to allow
using local 'jiri init -fetch-optional' settings when checkout a
snapshot file. By default, when creating a snapshot file, jiri will
write its optional attributes settings to the snapshot file. When
another client is checking out this snapshot file, it will guarantee
the recreated workspace is identical to the one when snapshot is
created. When this new flag is set to true, jiri will ignore the
optional attributes from the snapshot file and use the local ones.

Bugs: VC-159
Tests: Local
Change-Id: Ic2ff086b826f06fba73177cef9a85558205d0885