Add support for making paths relative to cwd
Previously, all paths output by jiri grep were relative to jirix.Root,
which is rarely where I run jiri grep from. If I want to copy/paste a
path to open that file in my editor, I either need to do this in a
different terminal window that is at the jiri root, or I need to modify
the path.

This change adds a flag, -cwd-rel (defaulting off, so no default
behavior change) which makes the output paths relative to the current
directory. This allows me to easily select the matched path and paste it
in my current development context, which is my preferred mode of
operation.

If a user wishes to default to this behavior, that may be achieved by
creating a shell function like:

	function jgrep() {
		jiri grep -cwd-rel "$@"
	}

Change-Id: I8751bf0f8148f0088d35082f9fa8420ebea6c43a