[analytics] Fix analytics test flake.
Before this patch, AnalyticsObject function will only check the 'second'
in time.Time struct to determine if the struct is not initialized. In
this case, there is a 1/60 chance that the result is false positive if
the Time struct is in fact initialized at L80. This patch uses
'Time.IsZero()' function to fix this issue.

Test: Local and CQ
Change-Id: I42070c0003400feae2b67d4b37416cb33b56bda6