[jiri] Revise 'rebase in progress' detection.
Due to recent `keep-empty` change, the git rebase creates rebase-merge
file instead of rebase-apply file when rebase is in progress, which
makes jiri failed to do rebase-abort when rebase failed, making repo
in a weird state. This change revise the 'rebase in progress' detection
in git package to fix this issue.

Bug: 48605
Change-Id: I09a23e2c095b448aa9bc1055625e967483277c8e