[githook] remove post-commit hook
The post-commit hook served only to warn about detached-HEAD mode, which
Git does natively in two places[1][2], effectively warning every time the
user enters or leaves detached-HEAD mode.

It seems post-commit is also ill-defined during rebase - depending on the
backend, it may or may not be called for new commits during rebase. See [3]
and the thread of replies to it for more details.

Bug: 48910
Change-Id: Ife3ff11fce1ed04f60f32b7f91bedb4780a5ca17

[1] https://github.com/git/git/blob/076cbdcd739aeb33c1be87b73aebae5e43d7bcc5/builtin/clone.c#L780
[2] https://github.com/git/git/blob/51ebf55b9309824346a6589c9f3b130c6f371b8f/builtin/checkout.c#L878
[3] https://lore.kernel.org/git/20200110231436.GA24315@google.com

Change-Id: Ife3ff11fce1ed04f60f32b7f91bedb4780a5ca17