[patch] Add special retcode for missing project
If a build fails to run `jiri patch` because the specified project is
missing, it will currently turn purple. I think this is most likely to
happen when someone tries to run a tryjob on a CL in a project that's
not included in the tryjob's checkout (example:
https://ci.chromium.org/b/8861463788089455632)

Use a designated retcode for such failures so the recipes can
disambiguate them from other patch failures, by turning red instead of
purple and showing a better message in the build summary.

Bug: 66079
Change-Id: I16f9175289d6fc7f7f2f0e074b23f2d9100364f4
Reviewed-on: https://fuchsia-review.googlesource.com/c/jiri/+/459254
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Commit-Queue: Oliver Newman <olivernewman@google.com>